refactor: split the inferenceEngine into separate files by functionality
