+++ src/inference/engine/index.ts
@@ -1,0 +1,6 @@
+export { ASTProcessor } from "./astProcessor";
+export { BinaryOpInference } from "./binaryOpInference";
+export { CollectionInference } from "./collectionInference";
+export { DeclarationInference } from "./declarationInference";
+export { ExpressionInference } from "./expressionInference";
+export { TypeParser } from "./typeParser";
